+++ src/Sulu/Bundle/AdminBundle/FormMetadata/FormMetadata.php
@@ -15,6 +15,8 @@ use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\SchemaMetadata;
 use Sulu\Component\Content\Metadata\PropertiesMetadata;
 
 /**
+ * @deprecated use Metadata\FormMetadata\FormMetadata instead.
+ *
  * Represents metadata for a form structure.
  */
 class FormMetadata extends PropertiesMetadata
