+++ src/Sulu/Bundle/AdminBundle/Resources/config/services.xml
@@ -43,24 +43,23 @@
             class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FormMetadataProvider"
         >
             <argument type="service" id="sulu_admin.form_metadata.form_xml_loader"/>
-            <argument type="service" id="sulu_page.structure.factory"/>
+            <argument type="service" id="sulu_admin.form_metadata.structure_loader"/>
             <argument type="service" id="sulu_core.expression_language" />
-            <argument>%sulu_core.locales%</argument>
             <argument>%sulu_admin.forms.directories%</argument>
             <argument>%sulu.cache_dir%/forms</argument>
             <argument>%kernel.debug%</argument>
-            <argument type="tagged" tag="sulu_admin.metadata_loader"></argument>
+            <argument type="tagged" tag="sulu_admin.form_metadata_loader"></argument>
             <tag name="sulu_admin.metadata_provider" type="form" />
             <tag name="kernel.cache_warmer" />
         </service>
 
         <service
-            id="sulu_admin.form_metadata_loader"
+            id="sulu_admin.form_metadata_xml_loader"
             class="Sulu\Bundle\AdminBundle\Metadata\FormMetadata\FormMetadataXmlLoader"
         >
             <argument type="service" id="sulu_admin.form_metadata.form_xml_loader"/>
             <argument>%sulu_admin.forms.directories%</argument>
-            <tag name="sulu_admin.metadata_loader"/>
+            <tag name="sulu_admin.form_metadata_loader"/>
         </service>
 
         <service
@@ -81,13 +80,21 @@
         <service id="sulu_admin.form_metadata.form_xml_loader" class="Sulu\Bundle\AdminBundle\FormMetadata\FormXmlLoader">
             <argument type="service" id="sulu_page.structure.properties_xml_parser"/>
             <argument type="service" id="sulu_page.structure.schema_xml_parser"/>
+            <argument>%sulu_core.locales%</argument>
+            <argument type="service" id="sulu_admin.form_metadata.form_metadata_mapper"/>
         </service>
 
         <service
-            id="sulu_admin.properties_metadata.properties_xml_loader"
-            class="Sulu\Bundle\AdminBundle\Metadata\PropertiesMetadata\PropertiesXmlLoader"
-        >
-            <argument type="service" id="sulu_page.structure.properties_xml_parser"/>
+            id="sulu_admin.form_metadata.form_metadata_mapper"
+            class="Sulu\Bundle\AdminBundle\FormMetadata\FormMetadataMapper">
+            <argument>%sulu.cache_dir%/forms</argument>
+            <argument>%kernel.debug%</argument>
+        </service>
+
+        <service id="sulu_admin.form_metadata.structure_loader" class="Sulu\Bundle\AdminBundle\FormMetadata\StructureLoader">
+            <argument type="service" id="sulu_page.structure.factory"/>
+            <argument type="service" id="sulu_admin.form_metadata.form_metadata_mapper"/>
+            <argument>%sulu_core.locales%</argument>
         </service>
 
         <service id="sulu_admin.route_registry" class="Sulu\Bundle\AdminBundle\Admin\RouteRegistry">
